[Assignment2] Convert to a plugin.
Credit to:
Myles Borins
TheAlphaNerd
https://github.com/TheAlphaNerd
This commit is modeled after his assignment 2 code below:
https://github.com/TheAlphaNerd/hueniversity/commit/c0d9d04723922632ac56be213609d9ee11e8bd6a

I learned a lot from his code.
The distributed classroom is working :-)
