+++ lib/index.js
@@ -1,30 +1,25 @@
+// ./lib/index.js 
 // Load modules
 
 var Hapi = require('hapi');
 var Hoek = require('hoek');
 var Package = require('../package.json');
 
-
 // Declare internals
-
-var internals = {};
+var internals = {
+    version: require('./version')
+};
 
 
 internals.init = function () {
 
     var server = new Hapi.Server();
+
     server.connection({ port: 8000 });
 
-    server.route({
-        method: 'GET',
-        path: '/version',
-        config: {
-            description: 'Returns the version of the server',
-            handler: function (request, reply) {
-
-                return reply({ version: Package.version });
-            }
-        }
+    server.register(internals.version, function (err) {
+		
+        Hoek.assert(!err, err);
     });
 
     server.start(function (err) {
@@ -33,5 +28,6 @@ internals.init = function () {
         console.log('Server started at: ' + server.info.uri);
     });
 };
+
 
 internals.init();
