+++ lib/version.js
@@ -1,0 +1,39 @@
+/*
+ * ./lib/version.js
+ * Script modelled after TheAlphaNerd's assignment2 code.
+ * https://github.com/TheAlphaNerd
+ */ 
+
+
+var internals = {
+        response: {
+            version: require('../package.json').version
+        }
+};
+
+
+// Register plugin
+exports.register = function (server, options, next) {
+
+    server.route({
+        method: 'GET',
+        path: '/version',
+        config: {
+            description: 'Returns the version of the server',
+		    handler: function (request, reply) {
+			
+                return reply(internals.response);
+            }
+        }
+    });
+	
+    return next();
+};
+
+
+// Plugin metadata
+exports.register.attributes = {
+    name: 'version'
+};
+
+
