+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "hueniversity",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "Community learning experiment",
   "main": "lib/index.js",
   "repository": {
